Json_reader for list_and_option and arithmetic using different strategy
Description: I've written manually both json_reader, but it should not
be hard to autogenerate them from grammar.json. The purpose of this diff
is to show a different strategy for the code generation of the json_reader.
In this strategy I don't think we need to return a list of intermediates
in codegen_types. Moreover it handles list and option correctly, which I
think the codegen of ulzii could not handle. In fact running
the -codegen_json_reader on list_and_option/grammar.json results in errors.
See the json_reader_pb.re.

Test plan:
make
make test