Allow specifying the headers for treesitter via the environment
Similar to TREESITTER_LIBDIR, the headers for treesitter may be
installed in a non-standard location. semgrep's `make configure` will
correctly set the environment variables, but `dune build` wasn't
running cc with the right include paths, so `#include
<tree_sitter/api.h>` wasn't resolving

I tested this by applying this patch to all of the generated
submodules (using a Very Advanced sed script).