fix: Strip trailing spaces in pkg-config output
On Fedora 36, pkgconf (https://github.com/pkgconf/pkgconf) ships as a
replacement for pkg-config. pkg-config does not include a trailing space
in the `--cflags-only-I` and `--libs-only-I` output, whereas pkgconf
does. This leads to linking errors when running `make test`:
https://gist.github.com/nmote/eb42fb763352f3980634763b38da9011

This PR works around the issue by stripping a trailing space from the
end of the `pkg-config` output, if it is present. This fixes the issue
on Fedora. I don't believe this will cause any issues on macOS, but I
don't have a way to test it.

Test plan:

```
$ ./configure
$ make test
```