Detect CC, AR, and STRIP to build tree-sitter on MinGW-w64
This is best-effort and should help picking the right cc, ar, and
strip utilities, whether they are prefixed (as when using the
MinGW-w64 cross-compiler) or not.

For a refresh of the shell features used, see:
POSIX 2018 2.6.2 Parameter Expansion.
https://pubs.opengroup.org/onlinepubs/9799919799/utilities/V3_chap02.html#tag_19_06_02

This includes an update the Makefile patch to build with MinGW-w64.

GNU Make 4.4.1 build for x86_64-pc-cygwin is perfectly capable of
reading the Makefile. It needs to be taught how to use MinGW-w64 to
build a DLL. This is a backport of patches accepted upstream.

See https://github.com/tree-sitter/tree-sitter/pull/4201 for the
corresponding upstream fix.

Co-authored-by: Antonin Décimo <antonin@tarides.com>